Moved all (?) tests to gym/tests
