--- gym/envs/unittest/__init__.py
@@ -1,4 +1,0 @@
-from gym.envs.unittest.cube_crash import CubeCrash
-from gym.envs.unittest.cube_crash import CubeCrashSparse
-from gym.envs.unittest.cube_crash import CubeCrashScreenBecomesBlack
-from gym.envs.unittest.memorize_digits import MemorizeDigits
