+++ tests/envs/robotics/hand/test_manipulate.py
@@ -3,13 +3,13 @@ import pickle
 import pytest
 
 from gym import envs
-from gym.envs.tests.spec_list import skip_mujoco, SKIP_MUJOCO_WARNING_MESSAGE
+from tests.envs.spec_list import skip_mujoco, SKIP_MUJOCO_WARNING_MESSAGE
 
 
 ENVIRONMENT_IDS = (
-    "HandManipulateEggTouchSensors-v1",
-    "HandManipulatePenTouchSensors-v0",
-    "HandManipulateBlockTouchSensors-v0",
+    "HandManipulateEgg-v0",
+    "HandManipulatePen-v0",
+    "HandManipulateBlock-v0",
 )
