+++ tests/vector/test_shared_memory.py
@@ -6,10 +6,10 @@ from multiprocessing.sharedctypes import SynchronizedArray
 from multiprocessing import Array, Process
 from collections import OrderedDict
 
-from gym.spaces import Box, Tuple, Dict
+from gym.spaces import Tuple, Dict
 from gym.error import CustomSpaceError
 from gym.vector.utils.spaces import _BaseGymSpaces
-from gym.vector.tests.utils import spaces, custom_spaces
+from tests.vector.utils import spaces, custom_spaces
 
 from gym.vector.utils.shared_memory import (
     create_shared_memory,
