+++ tests/vector/test_vector_env_wrapper.py
