+++ tests/wrappers/monitoring/test_video_recorder.py
