+++ tests/wrappers/test_gray_scale_observation.py
