+++ tests/wrappers/test_pixel_observation.py
